+++ CinnamonMagicLamp@klangman/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.0.2
+
+* Removed the need to change the System-Settings->Effect settings to "none" for the Minimize/Unminimize options. This is now accomplished by intercepting a cinnamon API and forcing it to disable the Cinnamon Minimize/Unminimize effects while the MagicLamp extension is enabled.
+* Fixed some minor typos in the README.
+
 ## 1.0.1
 
 * Initial version committed to cinnamon spices